feat: add recogize or get face api

+++ server/schemas/face.go
@@ -12,6 +12,6 @@ type FaceFindReq struct {
 }
 
 type FindFaceResp struct {
-	ID   uint   `json:"id,omitempty"`
+	ID   int64  `json:"id,omitempty"`
 	Name string `json:"name,omitempty"`
 }
